simplify text processing code and update notebook
