+++ examples/tacotron2_aishell3/chinese_g2p.py
@@ -13,9 +13,17 @@
 # limitations under the License.
 
 from typing import List, Tuple
+from pypinyin import lazy_pinyin, Style
+from preprocess_transcription import split_syllable
 
-from chinese_text_to_pinyin import convert_to_pinyin
-from chinese_phonology import split_syllable
+
+def convert_to_pinyin(text: str) -> List[str]:
+    """convert text into list of syllables, other characters that are not chinese, thus
+    cannot be converted to pinyin are splited.
+    """
+    syllables = lazy_pinyin(
+        text, style=Style.TONE3, neutral_tone_with_five=True)
+    return syllables
 
 
 def convert_sentence(text: str) -> List[Tuple[str]]:
